Move clear signed selectors to evm-tools lib
