+++ libs/evm-tools/README.md
@@ -2,4 +2,4 @@
 
 ### Description
 
-This library is in charge of packaging a set of tools used with EVMs need by coin integrations, Nano app bindings & UI clients.
+This library is in charge of packaging a set of EVM related tools, needed by coin integrations, Nano app bindings & UI clients.
