+++ libs/evm-tools/src/index.ts
@@ -1,3 +1,4 @@
 /* istanbul ignore file: pure exports, test unecessary */
 
 export * from "./message";
+export * from "./selectors";
